+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 * **AlertService** new service providing functionality to manage alerts. Those can also be used in a plugin.
 * **terra-alert** is now deprecated. Use new `AlertService` instead.
+* **ibanValidator** new validator for reactive forms checking whether a text is a valid IBAN.
+* **TerraValidators** new class that provides a set of additional validators for reactive forms.
 
 <a name="3.1.0"></a>
 # 3.1.0 (07.01.2019)